maybe make BUILD file header checking work again? lol
